remove output, use distribution instead of random

+++ RayTracingWeekend/hitable_list.h
@@ -10,7 +10,7 @@ public:
 	{
 		hit_record temp_rec;
 		bool hit_anything = false;
-		double closet_so_far = t_max;
+		float closet_so_far = t_max;
 		for (int i = 0; i < list_size; i++)
 		{
 			if (list[i]->hit(r, t_min, closet_so_far, temp_rec))
